fastapi/requirements.txt
fastapi[standard]
pydantic
pyodbc
cryptography
mysql-connector-python